pyproject.toml
[tool.poetry]
name = "nba-extractor"
version = "0.1.0"
description = ""
authors = ["Davi Vinicius Bogo <bogodavi10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.10.12"
nba-api = "^1.7"
pandas = "^2.0.0"
apache-airflow = "^2.10.1"
google-cloud-dataform = "^0.5.14"
pydantic-settings = "^2.7.1"
pandas-gbq = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
